Dehghani_Roya
This file is an API that was written using flask python. Its input is an image and it returns JSON strings. This project was part of a bigger one that was written in Iran and introduced in Expo 2021 as one of the achievements of Iran and attract millions of dollars.